Movido atributo para classe certa
Movido atributo para classe certa
